v4/msa-db-dist/requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
coverage==6.3.2
execnet==1.9.0
Flask==2.0.3
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pymongo==4.0.1
pyparsing==3.0.7
pytest==7.0.1
pytest-flask==1.2.0
pytest-forked==1.4.0
pytest-ordering==0.6
pytest-repeat==0.9.1
pytest-xdist==2.5.0
python-dotenv==0.19.2
requests==2.27.1
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
zipp==3.8.0
